requirements.txt
robotframework==7.0
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6